fix(bridge): correct maxTxPrice type

+++ contracts/interfaces/bridge/Bridge.sol
@@ -15,7 +15,7 @@ interface IBridge {
         address indexed from,
         uint256 amount,
         uint256 tax,
-        uint256 maxTxPrice,
+        uint16 maxTxPrice,
         string receiver
     );
 
@@ -25,7 +25,7 @@ interface IBridge {
 
     event Refund(uint256 indexed id);
 
-    event RBF(uint256 indexed id, uint256 maxTxPrice);
+    event RBF(uint256 indexed id, uint16 maxTxPrice);
 
     event Paid(uint256 indexed id, bytes32 txid, uint32 txout, uint256 value);
 
@@ -45,7 +45,7 @@ interface IBridge {
         address sender;
         uint256 amount; // msg.value - tax
         uint256 tax; // tax for goat foundation
-        uint256 maxTxPrice;
+        uint16 maxTxPrice;
         uint256 updatedAt;
         string receiver;
         WithdrawalStatus status;
